Starting over - again
When I created this repo I gave it the same name as the existing project folder because I assumed it would simply add its new files to the folder and adopt the whole folder. What it did instead was create the repo directory on top of the existing dir, overwriting and it deleting all of my files.
So I'm starting from 0 for the third time.