This commit is made mid-fix for track reversing
Need to switch tasks for a little bit, so this isn't a completed fix.
The plan is to, right before switching, sample the destination point's rotational basis and compare it to the train-body's - then if they are fall outside a range of similarity, reverse the sequence of the track's curve points - reversing the track's "forward" orientation - before reparenting the train to the nearest offset.