custom signals ?????
Extremely lost in the sauce, it is now too late and I am too sleepy to even keep track of what it is I'm trying to do and how.
Reversing all the points in a curve wasn't working - I mean it was more or less working, but it was hitching the system. So new idea: Just flip the train around and reverse the directional controls, so it *appears* to maintain its facing. Can barely even get through typing THAT sentence.
Feels like I'm drifting farther and farther away from the goal.